# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      id:
        format: int64
        type: integer
      name:
        maxLength: 128
        minLength: 3
        type: string
      password:
        maxLength: 128
        minLength: 4
        type: string
      phone:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/models.ValidationErrorResponse'
        type: array
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 128
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        maxLength: 128
        minLength: 3
        type: string
      password:
        maxLength: 128
        minLength: 4
        type: string
      phone:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.User:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      deleted_at:
        description: Дата удаления
        type: string
      email:
        description: Email
        type: string
      id:
        type: integer
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: ФИО
        type: string
      password:
        description: Пароль
        type: string
      phone:
        description: Телефон
        type: string
      status:
        description: Статус
        type: integer
      updated_at:
        description: Дата последнего обновления
        type: string
    type: object
  models.UserView:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      email:
        description: Email
        type: string
      id:
        type: integer
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: ФИО
        type: string
      phone:
        description: Телефон
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Роль
      status:
        description: Статус
        type: integer
    type: object
  models.ValidationErrorResponse:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Приложения бронирования спортивных площадок.
  title: Спортивный город
  version: "1.0"
paths:
...
  /api/auth:
    post:
      description: Создание нового пользователя
      parameters:
      - description: Данные для создания пользователя
        in: body
        name: createUser
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Создание нового пользователя
      tags:
      - Пользователи
  /api/auth/info:
    get:
      consumes:
      - application/json
      description: Получение информации о текущем аутентифицированном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Возвращает информацию о текущем пользователе
      tags:
      - Аутентификация
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по email и паролю с возвратом JWT токена
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT токен
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Аутентификация
  /api/auth/refresh:
    post:
      description: Обновление JWT токена для аутентифицированного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Новый JWT токен
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление JWT токена
      tags:
      - Аутентификация
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Получение списка всех пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            type: Bad
        "500":
          description: Internal Server Error
          schema:
            type: Internal
      summary: Возвращает список всех пользователей
      tags:
      - Пользователи
    put:
      description: Обновление существующего пользователя
      parameters:
      - description: Данные для обновления пользователя
        in: body
        name: updateUser
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: "No"
        "404":
          description: Not Found
          schema:
            type: Not
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Обновление существующего пользователя
      tags:
      - Пользователи
  /api/users/{id}:
    delete:
      description: Удаление пользователя по идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: "No"
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Удаляет пользователя по ID
      tags:
      - Пользователи
    get:
      description: Получение информации о пользователе по идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: Bad
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Возвращает информацию о пользователе по ID
      tags:
      - Пользователи
swagger: "2.0"